ci: rename AUR package to try auto cache busting
There's a recurring bug that when we publish a new release Arch users
report an error that the SHA sum is invalid. We think this is because,
even thought the PKGBUILD always downloads the new package, the
resulting filename remains the same. It seems likely that `pacman`
therefore doesn't see the file as new and so assumes that the _old_ SHA
sum is the official one.

So if we can make sure that the downloaded package always has a new
filename, then `pacman` should always then update the SHA sum associated
with it.